requirements.txt
alembic==1.2.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cryptography==3.2
entrypoints==0.3
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
flask-restful==0.3.8
idna==2.8
itsdangerous==1.1.0
jeepney==0.4.1
Jinja2==2.10.3
jsonconfig-tool==1.5
keyring==19.2.0
Mako==1.1.0
MarkupSafe==1.1.1
messages==0.5.0
pycparser==2.19
python-box==3.4.5
python-dateutil==2.8.0
python-decouple==3.1
python-editor==1.0.4
requests==2.22.0
SecretStorage==3.1.1
six==1.12.0
SQLAlchemy==1.3.10
urllib3==1.25.6
validus==0.3.0
Werkzeug==0.16.0
WTForms==2.2.1
